Fix Carol's last name
Carol generally renders her last name as "(Nichols || Goulding)". Follow that convention when mentioning the blog post announcing her as coauthor of "The Rust Programming Language".

Even if going with just the one last name, it would be "Nichols", not "Nicholas", but I think it's better to render her last name the way she prefers it.