Update 2020-10-28-this-week-in-rust.md
I wanted to learn how to use Rust (wasm modules) from JavaScript, but all the examples I could find either used webpack or where incomplete. Since I wanted to understand how this works without the added complexity of webpack (or any other bundler), it took a bit longer for me to find a solution. But the solution I found is as easy as it gets I think, only 1 html file, 1 js file, 1 rust file and 1 toml file is enough! So the link points to the GitHub Page I created which points to its own source and is hopefully a valuable resource for people how are looking for a Hello-World level, proof of concept of using Rust code in the browser.