add From Julia to Rust blog post
Hello Nell!

I was pointed to this newsletter by Rustacean friends. I hope this is the right way to add the blog post to the draft.
Thanks for working on this!

I also recently enjoyed your `Deep dive into the borrow checker` talk with the Devel labs group.
Cheers, and thanks for all your work with the Rust foundation.