Update 2021-10-27-this-week-in-rust.md
Here are links to a Rust project that I'm currently running (a livestream series), including the latest video I've uploaded.

I'd also like to link to the associated GitHub repository. I've left it out for the moment in the interest of not spamming the list, but in future would it be permissible to link notable project updates from the week, such as the following:

* [eZanmoto/norpl: Add support for index ranges and `else if` statements](https://github.com/search?q=repo%3Aezanmoto%2Fnorpl%20committer-date%3A2021-10-13..2021-10-20&type=Commits)

Thanks!