Blackhat should be in "Projects/Tooling updates?"
I do think thats not the right fit for it, and must be in the **Rust Walkthroughs** sections, It's a project, for sure. But not such as a totally free project on rust (black hat - book), or even the a crate/tool? - Its kinda an article/how-to/tutorial too, maybe: Walktrought then?.