Adding the Brazilian unnoficial translation 🇧🇷
First things first, as always.
This is translation idea came from getting people interested in The Rust programming language. And by that I mean there is no reason for me, to translate things like RFC's, English Events and others (as well jobs, by now).

This is just a small step for the rust community, but big step for the Brazilian's Rust Community. :)  

The right way maybe should be "PT-BR" - but I'll just change to this when Portugal give us our gold and "Pau Brasil".