add a tool for detecting unrecognized html tags
The markdown syntax allows un-escaped < > characters to be emitted in
the html output. This is probably undesirable, as something like Vec<T>
will be interpreted as a tag.

This file imports some code from inspect_links, which might be a bad
design; separating "CLI tool" from "library" code is left for a future
commit.