create_draft: auto-compute directory paths
Instead of requiring the user to run this script from a specific
directory, compute the draft path from the location of the running
script. This means if you run the script from the tools/ directory (or
anywhere else), it will do the right thing.

Adds a commandline option to override the default path.