create_draft: copy events from previous issue
The create_draft script will now search for a previous issue (under
draft and content) and then search that file for a range of lines that
look like the events list, and copy those lines over to the new draft
file.

If it has trouble finding the previous issue file, or can't find the
events list, it will print a warning and then leave a placeholder
comment in the draft issue where the events list should go.