Parity Jobs + Education program information
If the description about the education Program is too long, perhaps we can add that blurb to the miscellaneous section? Or alternatively, use the following text?

[Rust / Blockchain Academy with Parity (Expenses paid, on-site, Cambridge, July 11-August7)](https://docs.google.com/forms/d/e/1FAIpQLScAFegoJeHQDMXI0RJMI4OsbV0N2aXS5WmoaEcy5ysbaT0FCw/viewform)


Also, would be happy to connect about the program and send more details if you think the community would be interested; it's a unique program and a pilot effort to evangelize rust and blockchain from Parity.