Add cargo-semver-checks v0.15.0 release notes
The 0.15.0 release of cargo-semver-checks contains a massive number of improvements. I would like to get your thoughts on whether including the release notes in the "Project/Tooling Updates" section would be an appropriate way to communicate this to the TWIR readers, or if there's something else you'd recommend instead.

Thanks again for all the awesome work you're doing!