add search
largely based off instruction in their readme: https://github.com/pelican-plugins/search#installation

considered using tipue_search but was marked as deprecated in the repo and suggested the search plugin

this search also uses stork, which is written in rust!

also just adding it to the main page, since i'm not sure if it'd be as useful on every article or any other page.