Zenoh 0.7.0 Release -- Rust Pub/Sub/Query Protocol
As commented on the rejected pull request, Zenoh is a 100% rust protocol (See https://crates.io/crates/zenoh and http://github.com/eclipse-zenoh/zenoh). We have had the releases covered in the past (see content/2022-10-19-this-week-in-rust.md), not sure why the pull request for the last instalment was rejected, thus re-iterating.