Self-nominate "This Week in Ars Militaris #4" to 2023-08-02
Hi! I've added the latest issue of This Week in Ars Militaris, the weekly report about its development.

Ars Militaris is a strategy game set in the Ancient Rome Era, being made with Bevy Engine.

This issue (and the past ones), relate the experience of using Bevy and of contributing back to the open-source community for the first time (our first PR!).

It is a commercial game, but we're currently studying a licensing model that will be in according to our business model, with the goal of making our game open-source.