Self-nominate This Week in Ars Militaris 5 to 2023-08-09 (#4606)
Hi! I've added the latest issue of This Week in Ars Militaris, the weekly report about its development.

Ars Militaris is a strategy game set in the Ancient Rome Era, being made with Bevy Engine.

This issue contains the 0.1 release of our game, as well as the announcement that we're open-sourcing it.

Last time we submitted our issue to This Week in Rust, we were suggested to have a dedicated landing page for the article in question. Now, we ask if a "anchor" link to the article in the main page is enough (like we did in the commit, we would prefer this as we like that our website is a single page for now), or if it is really necessary to have a dedicated landing page. If it is, then please change the link we submitted to be this one (that contains the landing page):

https://arsmilitaris.com/this-week-in-ars-militaris-5.html

Also, we submit this issue to the "Project/Tooling Updates" section, because, even though This Week in Ars Militaris is also a newsletter, it is a newsletter about a specific project. But, if in your understanding it would fit better into the "Newsletters" section, feel free to include it there.

We deeply appreciate having our last issue accepted into This Week in Rust, and look forward for the feedback for this issue.

Co-authored-by: Brad Gibson <u007d@humanenginuity.com>