Added most recent episode in a series
I'm not sure if I should have included a link to the playlist with the rest of the episodes so I did and figured it was easier to remove than for the editor to find. I also wasn't sure if I should used both "tags" video and series so I did and again easier to remove than add.