Self-nominate This Week in Ars Militaris VI to 2023-08-16-this-week-in-rust.md (#4624)
Hey! We submit This Week in Ars Militaris VI to the this issue of This Week in Rust.

Ars Militaris is a commercial and open-source strategy game set in the Ancient Rome era, being built with Bevy Engine.

This week's issue contains several pieces that can be interesting to the Rust community, alongside other news related to Ars Militaris.

We have decided to start including This Week in Ars Militaris in the "Newsletters" section instead of the "Project/Tooling" section, as we now feel it is more appropriate for our newsletter.

As usual, we are deeply grateful for the inclusion of the past issues of This Week in Ars Militaris into This Week in Rust, and are open to feedback on this issue.