Add content for this week I came across (#4619)
* Rust videos for this week

* Remove duplicate

* AWS Podcast episode that covered rust for AI

The podcast in general is not only about rust but this episode was about rust and covered how rust is making in roads for use in AI.

* Change to Goolge Podcast Link

I couldn't find a more suitable link. The previous one worked fine in incognito but asked me to login if I just clicked it normally. This one doesn't ask me to login in either case.

* Add new video

* Add blog post

* Add video on converting websites to rust for frontend

* Added most recent episode in a series

I'm not sure if I should have included a link to the playlist with the rest of the episodes so I did and figured it was easier to remove than for the editor to find. I also wasn't sure if I should used both "tags" video and series so I did and again easier to remove than add.