Self-nominate This Week in Ars Militaris VII to 2023-08-23-this-week-in-rust.md
Hello! We add issue VII of This Week in Ars Militaris to the Newsletter section.

This week's issue contains some content that can be relevant to the Rust community, namely, in the "MacOS Release" section, along with more Ars Militaris-specific content.

We've also started offering lessons on Bevy targeted at beginners.

As usual, we're grateful for the inclusion of This Week in Ars Militaris in This Week in Rust, and we're open to feedback on this week's issue.