Add annotated talk for "SemVer in Rust: Tooling, Breakage, and Edge Cases"
Some folks prefer written content instead of video, so over the weekend I wrote up my FOSDEM 2024 talk as a blog post. This allowed me to include some additional content that wasn't part of the talk — for example, some recent information from the weeks after I gave the talk.

The YouTube link with the talk video was in last week's TWiR, so to disambiguate this link whose title is similar, I appended "annotated talk." If you feel there's a better way to disambiguate, I'd be happy to edit this or have you edit it as you see fit.

Thanks for maintaining This Week in Rust!