Add event: in-person 10 years of Rust 1.0 celebration in Berlin, DE
Adding announcement for 2025-05-15 19:00:

== Celebrating 10 years of Rust 1.0 ==

We are celebrating the 10 years anniversary of Rust 1.0!

In this event, we are going to meet, eat,
and attend the [international online celebration][1]
together with many other local Rust interest groups.

Everybody is welcome! No matter if you are very
experienced in Rust or have never had contact with
it before, please come join in and celebrate with us!

[1]: https://berline.rs/2025/05/15/celebrating-10-years-of-rust-10.html