Observations/Thoughts: Are we Teaching Rust Effectively?
I wanted to make a broad suggestion for improving the quality of learning materials
especially for addressing the perception of Rust's learning curve. It is unclear
what would make quantifiable improvements, but in this article I explore the possibility
that focusing on ownership may not be the most effective way to help students become
proficient.

A blog post seemed like the best first step, since this is informal opinion, at best.